requirements.txt
discord.py~=1.5.0
JustWatch~=0.5.0
discord-interactions==0.1.2
Flask~=1.1.0
